split NEF
