+++ packages/neon-core/src/sc/CallFlags.ts
@@ -1,0 +1,12 @@
+export enum CallFlags {
+  None = 0,
+  ReadStates = 0b00000001,
+  WriteStates = 0b00000010,
+  AllowCall = 0b00000100,
+  AllowNotify = 0b00001000,
+  States = ReadStates | WriteStates,
+  ReadOnly = ReadStates | AllowCall,
+  All = States | AllowCall | AllowNotify,
+}
+
+export default CallFlags;
