+++ packages/neon-core/src/sc/index.ts
@@ -11,4 +11,6 @@ export * from "./OpToken";
 export * from "./contracts";
 export * from "./fees";
 export * from "./NEF";
+export * from "./CallFlags";
+export * from "./MethodToken";
 export * from "./types";
